refactor: split admin create event form
